requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# Vector Database (lightweight alternative to Milvus)
faiss-cpu>=1.7.4

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web Dashboard
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
python-dateutil>=2.8.2
tqdm>=4.65.0
pyyaml>=6.0
colorama>=0.4.6

# Logging
structlog>=23.1.0

# Testing
pytest>=7.4.0
